% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.visual.R
\docType{package}
\name{spectra.visual-package}
\alias{spectra.visual}
\alias{spectra.visual-package}
\title{spectra.visual: a photoreceptor absorbance and ocular transmission database}
\description{
spectra.visual provies a repository of receptor absorbance and ocular transmission
spectra and metadata in a convenient format, along with simple search tools to aid
in data retrieval and management.
}
\section{Package contents}{

The package currently contains 281 spectra from 81 species.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomased/spectra.visual/}
  \item Report bugs at \url{https://github.com/thomased/spectra.visual/issues}
}

}
\author{
\strong{Maintainer}: Thomas E. White \email{thomas.white026@gmail.com} (0000-0002-3976-1734)

Authors:
\itemize{
  \item Hugo Gruson \email{hugo.gruson+R@normalesup.org} (0000-0002-4094-1476)
  \item Rafael Maia \email{rm72@zips.uakron.edu} (0000-0002-7563-9795)
}

}
