% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specvis_add.R
\name{specvis_add}
\alias{specvis_add}
\title{Add modelled sensitivity data}
\usage{
specvis_add(data, lmax, genus, species, citation_abbrev, doi, method,
  modelled = TRUE, filtered = FALSE, sub = NA, sex = NA,
  note = NA)
}
\description{
Temporary helper function while I'm adding modelled sensitivity data to the database. Not for release.
}
\author{
Thomas White \email{thomas.white026@gmail.com}
}
\keyword{internal}
