% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specvis_dict.R
\docType{data}
\name{specvis_dict}
\alias{specvis_dict}
\title{Dictionary for metadata}
\description{
Dictionary specifying the definitions of metadata.
}
\author{
Thomas White \email{thomas.white026@gmail.com}
}
\keyword{data}
