% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specvis_rebuild.R
\name{specvis_rebuild}
\alias{specvis_rebuild}
\title{Rebuild the database}
\usage{
specvis_rebuild()
}
\description{
Helper function to rebuild the database by saving raw data as compiled rda objects
}
\author{
Thomas White \email{thomas.white026@gmail.com}
}
\keyword{internal}
