% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specvis_search.R
\name{specvis_search}
\alias{specvis_search}
\title{Search the spectra.visual database}
\usage{
specvis_search(..., return.spectra = FALSE)
}
\arguments{
\item{...}{One or more search conditions to match (note: must use relational operators, e.g.
\code{==}, \code{<=}). Can match against any variables listed in the dictionary in
\code{\link{specvis_dict}}, namely:
\itemize{
 \item \code{spec_name}:
 \item \code{contributor}:
 \item \code{contributor_email}:
 \item \code{citation_abbrev}:
 \item \code{doi}:
 \item \code{method}:
 \item \code{type}:
 \item \code{modelled}:
 \item \code{genus}:
 \item \code{species}:
 \item \code{sub}:
 \item \code{sex}:
}}

\item{return.spectra}{A logical argument specifying whether the raw spectra be returned.
Defaults to \code{FALSE}, thereby returning all metadata associated with a query.}
}
\value{
A data frame of reflectance spectra from binned in 1 nm intervals
 from 300-700 nm, with the first column containing wavelengths ('wl').
}
\description{
Search the spectra.visual database.
}
\examples{
\dontrun{

# Query the database to see how many spectra from the genus 'bombus'
# are available, and examine their metadata
specvis_search(genus == "bombus")

# Save the corresponding spectra to an object
bombus_vis <- specvis_search(genus == "bombus", return.spectra = TRUE)
}

}
\author{
Thomas White \email{thomas.white026@gmail.com}
}
