% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specvis_spectra.R
\docType{data}
\name{specvis_spectra}
\alias{specvis_spectra}
\title{Photoreceptor absorbance and ocular transmission spectra}
\description{
Individual absorbance and transmission spectra, binned at 1 nm intervals from
 300 - 700 nm, beginning with a wavelength \code{'wl'} column.
}
\author{
Thomas White \email{thomas.white026@gmail.com}
}
\keyword{data}
